requirements.txt
streamlit~=1.28.2
langchain~=0.0.340
openai~=1.3.5
pinecone-client~=2.2.4
pandas~=1.5.3
python-dotenv~=1.0.0
tiktoken~=0.5.2
unstructured~=0.7.12
